Add changes for diamond configuration when support remote openstack db
When VSM is configured to connect remote openstack db, the diamond
configuration (/etc/diamond/diamond.conf) for the agent is also needed
to change. The "hostname" in the "[[VSMMySQLHandler]]" section
should be changed to the remote openstack db's ip
address, or else, the diamond service cannot connect to db successfully,
then the dashboard will not show the monitor data.